Implemented CSRF Protection when using mezzio-csrf package.
If the system uses mezzio-csrf and the request couldn't be validated, the Formhandler will pass the request with addition attribute "csrfError" to handle that error within your application.

As our Handler needs to know the field in which the CSRF Token is beeing submitted, you have to define that field as "type" => "csrf".